pyproject.toml
[tool.poetry]
name = "mtools"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
# readme = "README.md"
# package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.13.4"
ffmpeg-python = "^0.2.0"
mutagen = "^1.47.0"
utils-python = {git = "https://github.com/qwrwed/utils-python"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
